Move AKS to single recorded test class
