+++ tools/Azure.Mcp.Tools.Aks/tests/Azure.Mcp.Tools.Aks.Tests/assets.json
@@ -2,5 +2,5 @@
     "AssetsRepo": "Azure/azure-sdk-assets",
     "AssetsRepoPrefixPath": "",
     "TagPrefix": "Azure.Mcp.Tools.Aks.Tests",
-    "Tag": "Azure.Mcp.Tools.Aks.Tests_843f14df0d"
+    "Tag": "Azure.Mcp.Tools.Aks.Tests_35d7ac6287"
 }
